Laget JUnit test for hele JDBC og tilpasset JDBC klassen slik at den blir testbar. Ellers oppdaget og fikset enkelte logiske feil.

+++ src/GUI/BuddyController.java
@@ -107,7 +107,7 @@ public class BuddyController {
 			if ((boolean) link.isVisited()){
 				int linkID = database.getLinkID(link.getText());
 				int studID = database.getStudentID();
-				double rating = feedbackSlider.getValue();
+				int rating = (int) feedbackSlider.getValue();
 				database.insertFeedback(linkID, studID, rating);
 				link.setVisited(false);
 			}
